+++ Flow.Launcher.Plugin.ShortcutPlugin/Repositories/Interfaces/ICommandsRepository.cs
@@ -1,9 +1,8 @@
 ﻿using System.Collections.Generic;
-using Flow.Launcher.Plugin.ShortcutPlugin.Models.Shortcuts;
 
 namespace Flow.Launcher.Plugin.ShortcutPlugin.Repositories.Interfaces;
 
 public interface ICommandsRepository
 {
-    List<Result> ResolveCommand(List<string> arguments, string query);
+    List<Result> ResolveCommand(List<string> arguments, Query query);
 }