+++ Flow.Launcher.Plugin.ShortcutPlugin/Services/VariablesService.cs
@@ -1,11 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Flow.Launcher.Plugin.ShortcutPlugin.Extensions;
-using Flow.Launcher.Plugin.ShortcutPlugin.Repositories;
 using Flow.Launcher.Plugin.ShortcutPlugin.Repositories.Interfaces;
 using Flow.Launcher.Plugin.ShortcutPlugin.Services.Interfaces;
 using Flow.Launcher.Plugin.ShortcutPlugin.Utilities;
-using Flow.Launcher.Plugin.ShortcutPlugin.Utils;
 
 namespace Flow.Launcher.Plugin.ShortcutPlugin.Services;
 
@@ -25,14 +23,15 @@ public class VariablesService : IVariablesService
         if (variables.Count == 0)
             return ResultExtensions.EmptyResult("No variables found.");
 
-        return variables.Select(variable => new Result
-                        {
-                            Title = $"Variable '{variable.Name}'",
-                            SubTitle = $"Value: '{variable.Value}'",
-                            IcoPath = Icons.Logo,
-                            Action = _ => true
-                        })
-                        .ToList();
+        return variables
+            .Select(variable => new Result
+            {
+                Title = $"Variable '{variable.Name}'",
+                SubTitle = $"Value: '{variable.Value}'",
+                IcoPath = Icons.Logo,
+                Action = _ => true
+            })
+            .ToList();
     }
 
     public List<Result> GetVariable(string name)
@@ -44,7 +43,8 @@ public class VariablesService : IVariablesService
 
         return ResultExtensions.SingleResult(
             $"Variable '{variable.Name}'",
-            $"Value: '{variable.Value}'");
+            $"Value: '{variable.Value}'"
+        );
     }
 
     public List<Result> AddVariable(string name, string value)
@@ -52,7 +52,11 @@ public class VariablesService : IVariablesService
         return ResultExtensions.SingleResult(
             $"Add variable '{name}'",
             $"Value: '{value}'",
-            () => { _variablesRepository.AddVariable(name, value); });
+            () =>
+            {
+                _variablesRepository.AddVariable(name, value);
+            }
+        );
     }
 
     public List<Result> RemoveVariable(string name)
@@ -62,11 +66,14 @@ public class VariablesService : IVariablesService
         if (variable is null)
             return ResultExtensions.EmptyResult($"Variable '{name}' not found.");
 
-
         return ResultExtensions.SingleResult(
             $"Remove variable '{name}'",
             $"Value: '{variable.Value}'",
-            () => { _variablesRepository.RemoveVariable(name); });
+            () =>
+            {
+                _variablesRepository.RemoveVariable(name);
+            }
+        );
     }
 
     public List<Result> UpdateVariable(string name, string value)
@@ -79,7 +86,11 @@ public class VariablesService : IVariablesService
         return ResultExtensions.SingleResult(
             $"Update variable '{name}'",
             $"Old value '{variable.Value}' | New value '{value}'",
-            () => { _variablesRepository.UpdateVariable(name, value); });
+            () =>
+            {
+                _variablesRepository.UpdateVariable(name, value);
+            }
+        );
     }
 
     public void Reload()