+++ Flow.Launcher.Plugin.ShortcutPlugin/Utilities/Constants.cs
@@ -7,4 +7,7 @@ public static class Constants
     public const string VariablesFileName = "Config\\variables.json";
 
     public const string VariableFormat = "${{{0}}}";
+
+    public const string GithubRepository =
+        "https://github.com/mantasjasikenas/flow-launcher-shortcuts-plugin";
 }