+++ Flow.Launcher.Plugin.ShortcutPlugin/Utilities/IShortcutHandler.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using Flow.Launcher.Plugin.ShortcutPlugin.models;
 using Flow.Launcher.Plugin.ShortcutPlugin.Models.Shortcuts;
 
 namespace Flow.Launcher.Plugin.ShortcutPlugin.Utilities;