+++ Flow.Launcher.Plugin.ShortcutPlugin/Utilities/QueryBuilder.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Linq;
 
-namespace Flow.Launcher.Plugin.ShortcutPlugin.Utils;
+namespace Flow.Launcher.Plugin.ShortcutPlugin.Utilities;
 
 public static class QueryBuilder
 {
+    [Obsolete("Use Build instead")]
     public static Query Build(string rawQuery)
     {
         if (string.IsNullOrWhiteSpace(rawQuery))
@@ -18,11 +19,9 @@ public static class QueryBuilder
             return null;
         }
 
-
         var search = terms.ElementAtOrDefault(1);
         var searchTerms = terms[1..];
         var actionKeyword = terms.FirstOrDefault();
-
 
         var query = new Query(rawQuery, search, terms, searchTerms, actionKeyword);